requirements.txt
pdfplumber
python-docx
nltk
spacy
scikit-learn
streamlit
pandas
numpy